requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
gunicorn
requests
cloudinary
psycopg[binary]
